pyproject.toml
[project]
name = "lucid-t"
version = "0.1.0"
description = ""
authors = [
    {name = "Mechislav",email = "93447863+Mechislav1989@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "pydantic (>=2.11.1,<3.0.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "punq (==0.7.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "cachetools (>=5.5.2,<6.0.0)",
    "python-jose (>=3.4.0,<4.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "aiomysql (>=0.2.0,<0.3.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
